Fix #2 and #4 and #5.
This change makes the old package layout a bit less annoying.

Now we have:
- ghc
- ghc-dev
- cabal (replaces cabal-install and cabal-install-doc)
- stack

ghc has the most minimal package requirements as possible. Validated
this the hard way. Yes you need a compiler for ghci to work, so
gcc or llvm would be needed.

ghc-dev has basic packages needed to build against libraries the
compiler was built against. It also has the llc-3.5 and opt-3.5
binaries in the package. Not 100% sure this is the best option
or not but in the case of just running ghc/ghci they aren't
necessary.

Cabal updated to 1.22.7.0 as well. Stack is also now a static
binary. Size difference was 44.4MiB static vs 44.0MiB dynamic.

Finally, built ghc with LD=ld.gold. No real difference observed
and everything continues to build fine based on my testing.

Next up 8.0 rc work and nightly git builds.